feat: improve `/tasks/:id/assign`.
This changes the behavior of the endpoint if the task is linked to a delivery.
Instead of only assigning the task in question, the entire delivery is assigned.
This may be less compliant with REST principles, but leaving it as it was without
any validation would have allowed clients to assign a delivery to different riders.